feat: expose router for serving user interface
