+++ package.json
@@ -17,7 +17,7 @@
     "test": "jest",
     "setpackageversion": "tsc -p ./scripts/tsconfig.json && node ./scripts/setpackageversion.js",
     "verifypackagelock": "tsc -p ./scripts/tsconfig.json && node ./scripts/verifypackagelock.js",
-    "ui": "node ./lib/clUiRunner.js",
+    "ui": "node ./lib/uiServer.js",
     "commit": "git-cz",
     "commitmsg": "commitlint -E GIT_PARAMS"
   },
