--- src/clUiRunner.ts
@@ -1,10 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-/**
- * This is just a convenience for @conversationlearner/sdk developers to test the UI
- */
-import startCLUi from './clUi'
-
-startCLUi()
