+++ src/index.ts
@@ -7,12 +7,12 @@ import { ICLOptions } from './CLOptions'
 import { ClientMemoryManager, ReadOnlyClientMemoryManager } from './Memory/ClientMemoryManager'
 import { RedisStorage } from './RedisStorage'
 import { FileStorage } from './FileStorage'
-import startUiServer from './clUi'
+import uiRouter from './uiRouter'
 import { SessionEndState, MemoryValue } from '@conversationlearner/models'
 import { EntityDetectionCallback, OnSessionStartCallback, OnSessionEndCallback, LogicCallback, RenderCallback, ICallbackInput } from './CLRunner'
 
 export {
-    startUiServer,
+    uiRouter,
     ConversationLearner,
     ICLOptions,
     ClientMemoryManager,
